Fjern mdc korr på jobb
